moved major chunks into components and out of vounteer-onboarding page
